lift func to conftest
